[x64 Support]Fixed errors and warnings related to vgui

+++ src/vgui2/src/vgui.cpp
@@ -909,11 +909,11 @@ void CVGui::ShutdownMessage(unsigned int shutdownID)
 	VPANEL panel = g_pSurface->GetEmbeddedPanel();
 	for (int i = 0; i < ((VPanel *)panel)->GetChildCount(); i++)
 	{
-		g_pIVgui->PostMessage((VPANEL)((VPanel *)panel)->GetChild(i), new KeyValues("ShutdownRequest", "id", shutdownID), NULL);
+		g_pIVgui->PostMessage((VPANEL)((VPanel *)panel)->GetChild(i), new KeyValues("ShutdownRequest", "id", (int)shutdownID), NULL);
 	}
 
 	// post to the top level window as well
-	g_pIVgui->PostMessage(panel, new KeyValues("ShutdownRequest", "id", shutdownID), NULL);
+	g_pIVgui->PostMessage(panel, new KeyValues("ShutdownRequest", "id", (int)shutdownID), NULL);
 }
 
 //-----------------------------------------------------------------------------
